+++ backend/yigil-api/src/main/java/kr/co/yigil/follow/interfaces/controller/FollowApiController.java
@@ -5,8 +5,8 @@ import kr.co.yigil.auth.MemberOnly;
 import kr.co.yigil.auth.domain.Accessor;
 import kr.co.yigil.follow.application.FollowFacade;
 import kr.co.yigil.follow.domain.FollowInfo;
-import kr.co.yigil.follow.dto.response.FollowResponse;
-import kr.co.yigil.follow.dto.response.UnfollowResponse;
+import kr.co.yigil.follow.interfaces.dto.FollowResponse;
+import kr.co.yigil.follow.interfaces.dto.UnfollowResponse;
 import kr.co.yigil.follow.interfaces.dto.FollowDto.FollowersResponse;
 import kr.co.yigil.follow.interfaces.dto.FollowDto.FollowingsResponse;
 import kr.co.yigil.follow.interfaces.dto.FollowDtoMapper;
