+++ backend/yigil-api/src/main/java/kr/co/yigil/follow/interfaces/dto/FollowDto.java
@@ -46,4 +46,6 @@ public class FollowDto {
             private final boolean following;
     }
 
+
+
 }
