+++ backend/yigil-api/src/main/java/kr/co/yigil/follow/interfaces/dto/FollowResponse.java
@@ -1,6 +1,5 @@
-package kr.co.yigil.follow.dto.response;
+package kr.co.yigil.follow.interfaces.dto;
 
-import jakarta.persistence.NamedStoredProcedureQueries;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -8,7 +7,6 @@ import lombok.NoArgsConstructor;
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
-public class UnfollowResponse {
-
+public class FollowResponse {
     private String message;
 }
