+++ backend/yigil-api/src/main/java/kr/co/yigil/follow/interfaces/dto/UnfollowResponse.java
@@ -1,5 +1,6 @@
-package kr.co.yigil.follow.dto.response;
+package kr.co.yigil.follow.interfaces.dto;
 
+import jakarta.persistence.NamedStoredProcedureQueries;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -7,6 +8,7 @@ import lombok.NoArgsConstructor;
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
-public class FollowResponse {
+public class UnfollowResponse {
+
     private String message;
 }
